Use u32 as the block height uniformly

+++ src/btc/balance-history/src/btc/client.rs
@@ -6,11 +6,11 @@ pub trait BTCClient: Send + Sync {
     fn init(&self) -> Result<(), String>;
     fn stop(&self) -> Result<(), String>;
     
-    fn get_latest_block_height(&self) -> Result<u64, String>;
-    fn get_block_hash(&self, block_height: u64) -> Result<BlockHash, String>;
+    fn get_latest_block_height(&self) -> Result<u32, String>;
+    fn get_block_hash(&self, block_height: u32) -> Result<BlockHash, String>;
     fn get_block_by_hash(&self, block_hash: &BlockHash) -> Result<Block, String>;
-    fn get_block_by_height(&self, block_height: u64) -> Result<Block, String>;
-    async fn get_blocks(&self, start_height: u64, end_height: u64) -> Result<Vec<Block>, String>;
+    fn get_block_by_height(&self, block_height: u32) -> Result<Block, String>;
+    async fn get_blocks(&self, start_height: u32, end_height: u32) -> Result<Vec<Block>, String>;
     fn get_utxo(&self, outpoint: &OutPoint) -> Result<(ScriptBuf, Amount), String>;
 }
